// src/app/api/models/order-address-dto.ts
/* tslint:disable */
export interface OrderAddressDTO {
  landmark?: string;
  addressType?: string;
  city?: string;
  customerId?: string;
  houseNoOrBuildingName?: string;
  id?: number;
  alternatePhone?: number;
  name?: string;
  phone?: number;
  pincode?: number;
  roadNameAreaOrStreet?: string;
  state?: string;
}